backend/requirements.txt
Django==5.0
djangorestframework
psycopg2-binary
celery[redis]
django-celery-results
django-celery-beat==2.1.0
redis
django-environ
django-extensions
django-elasticsearch-dsl
django-cors-headers
ipdb
djangorestframework-simplejwt